requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via
    #   starlette
    #   watchfiles
build==1.2.2.post1
    # via pip-tools
click==8.1.8
    # via
    #   pip-tools
    #   uvicorn
colorama==0.4.6
    # via
    #   build
    #   click
    #   uvicorn
dnspython==2.7.0
    # via pymongo
fastapi==0.115.6
    # via -r requirements.in
h11==0.14.0
    # via uvicorn
httptools==0.6.4
    # via uvicorn
idna==3.10
    # via anyio
motor==3.6.1
    # via -r requirements.in
packaging==24.2
    # via build
pip-tools==7.4.1
    # via -r requirements.in
pydantic==2.10.5
    # via fastapi
pydantic-core==2.27.2
    # via pydantic
pymongo==4.9.2
    # via motor
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   uvicorn
pyyaml==6.0.2
    # via uvicorn
sniffio==1.3.1
    # via anyio
starlette==0.41.3
    # via fastapi
typing-extensions==4.12.2
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
uvicorn[standard]==0.34.0
    # via -r requirements.in
watchfiles==1.0.4
    # via uvicorn
websockets==14.2
    # via uvicorn
wheel==0.45.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
